*: Use &str instead of Into<String>
This prevents accidental allocations and copies within the library by
forcing the conversion to `String` to happen as late as possible. With
`Into<String>` parameters, arguments could be converted multiple times.

The only two types in libstd that can be passed as `Into<String>` are
`String` and `&str`. Accepting `&str` achieves the same, with the minor
change that a `String` must be passed by reference:

    let name: String = "my_metric".to_string();
    let help: String = "some help string".to_string();
    let opts = Opts::new(&name, &help);

In most uses of Prometheus, the strings will be `&'static str` anyway,
so this should not have much of an effect.