desc: Replace regex name validation with functions
This allows dropping the dependency on regex, making the instrumentation
library much more lightweight.

This also fixes a bug with metric name validation, where an initial
colon was incorrectly disallowed. See data model for correct format:
  https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels

This may be because the library closely follows the official Go library,
which also had this issue.

Refs #85
Refs https://github.com/prometheus/client_golang/pull/255