Add public export for MetricVec (#112)
CounterVec, GaugeVec, and HistogramVec are type aliases for MetricVec,
but since MetricVec isn't publicly exported, it's impossible to tell
from the docs what methods exist on MetricVec (but you can still call
them if you know what they are). This exports MetricVec publicly, so it
shows up properly in the docs.