Fix bug in global label handling.
There was a bug in the label handling where every metric after the first
did not have labels applied.

We were calling Vec::append() with a mutable vec as the argument, and
doing so causes the elements to change ownership from the source to the
target vec.

This change fixes the problem by cloning the label vec and introducing a
test.