static-metric/src/parser: Allow not specifying visibility
When determining whether to parse a struct or an enum the `Parse`
implementation for `StaticMetricMacroBodyItem` checks for the tokens
`struct` or `enum`. Given that the two can optionally be preceeded by
a visibility token (e.g. `pub(crate)`) one needs to peek at both the
first and the second token.

Signed-off-by: Max Inden <mail@max-inden.de>